read
Reading from terminal is so easy, gets, scanf, and many.
But no, we gotta get POSIX involved.

unistd.h is unix POSIX-1 interface, now available for
linux, bsd, and anything Paw-six compliant.

while is running for an empty body, because everything
is being done by the condition itself.

This is a red-flag. It should not have happened.
Even golang does this. WHY?

Do not generate new values in conditionals.
I consider even the expression checks too much in conditionals.
No wonder I do not and have not gotten done anything of value.
Such a waste of time and space.

Fret not. Low self-esteem is a hallmark of ADHD. Once you
know you can berate and humiliate yourself, and then realise
oh ADHD, gotta stop hurting self.

Anneway, read standard input by file no, which is 0 me thinks.
then pass ref to a local variable, and read only 1 byte questio.
and check whether we read one character.

close the reading by pressing ctrl+d or ctrl+c for exiting the
binary.

There are two modes the terminal has: canocial/cooked or raw.
We need to raw.
We need to strip away the cooked parts and reach the uncooked.
It is going to take time.