Docker: 2, Anubhav: 2
Oh man oh man; who ever needs docker. Docker is donkey stupid.

I was good with virtualbox, running it in headless mode,
getting shit done. No care of where it was running or who was
running it. Not caring what is even architecture. Everything is
x86_64 or amd64, amirice questio. Then came this horrible fruit
company that doctors hate and is apparently now a walled garden
exclaim. No, it was not content, had to poison humanity with newer
tech. Stupid newer processors do not run x86_64 that well,
do not allow virtualboxes to run properly if at all.

Fine, docker it is. Docker is donkey waalls. It sucks so much
oxygen out of the world just trying to figure out what the
pheck is even the configuartion for simpler things. Things that
were install in vm and forget have to be reinvented.
    For what, questio, questio, questio.

    I figured that I have been wrestling with docker on the
windows, may be try it back into the saner world i.e. linux.
Oh, boy, same two errros:
1. /usr/bin/ld: cannot open output file kilo.exe: Permission denied
2. mkdir: cannot create directory ‘build’: Permission denied
3. chmod: changing permissions of './entrypoint.sh': Operation not permitted
4. ...
5. ...

Three of those are results of fixing another. Actually begotten:
                                              as the saying goes,
                                   an error begets another three.

Anyway, fixed it. Read the changelog.

I am however questioning programming as a career.
    Maybe, I should pickup farming or teaching as a career,
        and coding as a hobby.
            Throw in statistically apt OnlyFans to make money
                on the side and I am swimming in Gold:
                    shower or not, you decide.

Toodles.